Give user ability to leave a review, and to view it in product details

+++ blog/views.py
@@ -1,12 +1,15 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import Review
 from django import forms
+from django.contrib import messages
 from .forms import ReviewForm
 from products.models import Product
+from .models import Rating
 
 
 def create_review(request, product_id):
     product = get_object_or_404(Product, pk=product_id)
+    rating = Rating.objects.filter(product=product, user=request.user).first()
     if request.method == 'POST':
         form = ReviewForm(request.POST)
         if form.is_valid():
@@ -14,12 +17,15 @@ def create_review(request, product_id):
             review.product = product
             review.user = request.user
             review.save()
-            return redirect('blog/create_review.html', pk=product_id)
+            messages.info(request, 'New review added')
+            return redirect('product_detail', product_id=product_id)
     else:
         form = ReviewForm()
 
     context = {
         'form': form,
+        'rating': rating,
+        'product': product,
     }
 
     return render(request, 'blog/create_review.html', context)
@@ -34,7 +40,7 @@ def view_reviews(request, product_id):
         'product': product,
         }
 
-    return render(request, 'blog/review.html', context)
+    return render(request, 'blog/reviews.html', context)
 
 
 def update_review(request, review_id):
@@ -50,7 +56,13 @@ def update_review(request, review_id):
 
 
 def delete_review(request, review_id):
+    """Delete a review"""
+    if not request.user.is_authenticated:
+        messages.error(request, "Sorry, \
+            you can't access other user's profiles.")
+        return redirect('product_detail', id=product_id)
+
     review = get_object_or_404(Review, id=review_id)
-    product_id = review.product.id
     review.delete()
-    return redirect('product_detail', id=product_id)
+    messages.info(request, 'Product deleted!')
+    return redirect(reverse('products'))
